Fix checkstyles failures
